#   Implement multi-processing in fastHadd to merge DQM histograms faster